allow compute nodes to be associated with host agg
The Compute API allows for a host_name attribute to be specified
along with the aggregate's ID to associate or disassociate that host
with the aggregate. Unfortunately, due to vestigial coupling between the
compute service record and the compute node concept, this meant that
Ironic baremetal resource nodes cannot effectively use the Nova host
aggregate concept to group resource nodes. Instead, only the
nova-compute service(s) that manage those Ironic resource nodes can be
grouped together using a Nova host aggregate. This is less than useful
for Ironic deployments.

This patch does not change the public Compute API for adding or removing
hosts from an aggregate. Instead, what this patch does is fall back to
looking up compute node records by hypervisor_hostname if a service
record with the supplied hostname parameter cannot be found.

This work should pave the way for Ironic deployments to fully take
advantage of Nova server group affinity/anti-affinity and allow
HostAggregate metadata filters to operate on Ironic baremetal nodes in
the scheduler.

Blueprint: aggregate-affinity

Change-Id: Ic7828846148fb365a6d25f97eff117ca4a47e0df

 Conflicts:
	nova/compute/api.py
	nova/tests/unit/compute/test_compute.py