Add stub unified limits driver
The unified limits driver is starting with the noop driver. This gives us
the closest API behaviour to what we describe in the spec.

The Unified Limits quota driver will has several purposes:

* stop all existing quota enforcement, so we can replace it
* stop checking the database for quota info
* make the API do what it does today with the noop driver enabled

The next few patches will assert the existing API behaviour with the
unified limits quota driver. This is the exact same thing that happens
today when you enable the noop driver.

As we add back limits, using the new unified limits approach, we will
update the API so users are informed about what limits are actually
being enforced.

blueprint unified-limits-nova

Change-Id: Iab152a6b2bb58454c32889390ec9add43771fa62