Assert quota related API behavior when noop
Adding tests so its clear what happens with the noop driver when using
the quota APIs.

To make the unit tests work, we had to make the caching of the quota
driver slightly more dynamic. We verify the current config matches
the currently cached driver, and reload the driver if there is a miss-match.
It also preserves the ability of some unit tests to pass in a fake
quota driver.

We also test the current unified limits driver, as it is currently
identical in behaviour to the noop driver.
As things evolve the tests will diverge, but will show the common
approach to what is returned from the API in both cases.

blueprint unified-limits-nova

Change-Id: If3c58d6cbf0a0aee62766c7142beab165c1fb9a4