Make unified limits APIs return reserved of 0
Currently the noop driver returns a reserved value of -1, and before
this patch the same is true with the unified limits driver.

Given the way the quota system currently works, the reserved value can
be ignored. With unified limits, we want to make the API look as
identical as possible to the DB driver, when you have the same limits
applied with the unified limits driver. As such, we change the API to
return a reserved value of 0 for the unified limits driver.

Longer term, when the API gets a new microversion to tidy up quotas, the
reserved value will likely be removed in the new microversion, because
it no longer has any real meaning.

blueprint unified-limits-nova

Change-Id: I28212857313ae72903d2139884750d5de690c6bd