Update quota sets APIs
Ensure the limit related APIs reflect the new reality of enforcing the
API and DB limits based on keystone only.

For now we skip all updates to the DB, as none mean anything to the new
code, as we only look at keystone now.

Note: this will need to be updated again once we add limits for
cores, ram, instances, etc.

blueprint unified-limits-nova

Change-Id: I5ef968395b4bdc6f190e239a19a723316b1d5baf