libvirt: Create qcow2 disks with the correct size without extending
All callers to the create_image method of the Qcow2 imagebackend provide
a size and so we should just always create the initial image using that
size. The name of the method is also misleading and changed to better
reflect that an image is being created here and not copied.

Change-Id: If2fa298c2e67266dcaf310e05bda51af4285526f