db: Add migration to resolve shadow table discrepancies
As part of the migration to alembic, a number of mistakes were noted in
the migrations. These were all a result of doing things differently for
the shadow table compared to the main table. Resolve some of these in a
new migration to both address the bugs and test out the alembic
machinery.

Note that we don't actually resolve all of the mismatches. A large
number of these mismatches were caused by the addition of
'nullable=False' to a column in a main table but not the equivalent
shadow table. With the benefit of hindsight, this is probably wise and
possibly even intended. If we were to apply the new constraint to the
shadow table also, we'd need a data migration to populate any remaining
NULL fields in the shadow tables. This data migration could be very
costly and may not even possible as we don't necessarily have all
available context required to do this population. As a result, the
wording around these differences is changed to indicate that we will
likely never address these "bugs".

Change-Id: Icdc683d534a4fd35af399561253888abb9193c7d
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>