Catch an exception in power off procedure
In libvirt driver, an exception of libvirt error may be thrown
when calling shutdown method.

This exception is caught by nobody, then the following retry
logic and the destroy method will be missed, causing the power
off procedure failure.

Catch this exception (as the retry logic does) to make the
procedure go on.

Change-Id: Icf9b15207a7e13333cf5640a5bce0e1082ecf4ec