Introduce remote_managed tag for PCI devs
PCI devices may be managed remotely from the perspective of a hypervisor
host (e.g. by a SmartNIC DPU) which means that the VF control plane is
not available to the hypervisor. Depending on the presence of a
remote_managed device attribute in the InstancePCIRequest spec and
available device types in a pool, additional processing needs to be
done:

* Filtering of devices marked as `remote_managed: "true"` in the
  whitelist configuration so that they are not used in legacy SR-IOV
  and hardware offload requests;

* Early error reporting if PFs marked as remote_managed="true" are
  present in the whitelist configuration. This is not supported
  explicitly since allocating such PFs would remove the associated
  VFs from the pool and an instance with such PF and its VFs will
  not have access to the control plane required for representor
  interface plugging at the SmartNIC DPU side. This configuration
  is not valid which is enforced in the PCIDeviceStats code.

* Checking of the presence of a card serial number in the PCI VPD
  capability of a device if it was marked as `remote_managed: "true"`
  in the whitelist. The card serial number presence is mandatory
  because it is used for identification of a host in the networking
  backend that will handle the configuration of a given PCI device at
  the remote host side (i.e. representor plugging, flow programming).

For compatibility, all devices not explicitly marked as remote_managed
in the whitelist are assumed to have remote_managed attribute set to
False.

Implements: blueprint integration-with-off-path-network-backends
Change-Id: Ic44d5e206326827d00a751da3cea67afe3929a08