imagebackend: default by_name image_type to config correctly
Callers of by_name shouldn't need to pass in the default from CONF.

This has implications for the libvirt_imagebackend fixture. The
_mock_backend() method in the libvirt_imagebackend fixture is what
mocks out Backend.backend(). Before this patch, the defaulting of
image_type to CONF.libvirt.images_type was done *inside*
Backend.backend() whenever image_type=None was passed in. As of this
patch, this is no longer case, and the default value from CONF gets
passed in instead.

Unit tests were testing the previous behaviour and essentially just
asserting that None was passed in. This was never really correct to
begin with, so a couple of assertIsNone() assertions need to be
updated.

Change-Id: Ib95aa440ffff0989c90f09f5b5d4e26b30db9683