Raise InstanceNotFound on fkey constraint fail saving info cache
We have seen tracebacks in the gate in the nova-multi-cell job during
cross-cell migrations of a failure to save an instance info cache due
to a foreign key constraint violation on instance_uuid. This can occur
if we happen to receive an external event from neutron after the
instance record has been moved to the destination cell.

This adds a catch of DBReferenceError from oslo.db and raises
InstanceNotFound if the key is instance_uuid, else it reraises.

One more step is required to prevent tracing which is to log at level
INFO instead of level ERROR in the case of InstanceNotFound. Then
reraise it for callers to handle.

Closes-Bug: #1944787

Change-Id: I71b7312879dab6857255c78c21ad8dd45d5ea00f