skip test_tagged_attachment in nova-next
This change adds
tempest.api.compute.servers.test_device_tagging.TaggedAttachmentsTest.test_tagged_attachment
to the tempest exclude regex

over the past few weeks we have noticed this test failing intermitently
and it has not started to become a gate blocker. This test is executed in other
jobs that use the PC machine type and is only failing in the nova-next
job which uses q35. As such while we work out how to address this properly
we skip it in the nova-next.

Conflicts in .zuul.yaml because 5d2f2da0afa changed the same
tempest_exclude_regex and is not in xena.

Change-Id: I845ca5989a8ad84d7c04971316fd892cd29cfe1f
Related-Bug: #1959899
(cherry picked from commit b00ce99dd456ab701cef0a9d4429920834d3d840)