libvirt: Add configuration options to set SPICE compression settings
This patch adds the following SPICE-related options to the 'spice'
configuration group of a Nova configuration:

  - image_compression
  - jpeg_compression
  - zlib_compression
  - playback_compression
  - streaming_mode

These configuration options can be used to enable and set the SPICE
compression settings for libvirt (QEMU/KVM) provisioned instances.
Each configuration option is optional and can be set explictly to
configure the associated SPICE compression setting for libvirt. If all
configuration options are not set, then none of the SPICE compression
settings will be configured for libvirt, which corresponds to the
behavior before this change. In this case, the built-in defaults from
the libvirt backend (e.g. QEMU) are used.

Note that those options are only taken into account if SPICE support is
enabled (and the VNC support is disabled).

Implements: blueprint nova-support-spice-compression-algorithm
Change-Id: Ia7efeb1b1a04504721e1a5bdd1b5fa7a87cdb810