Fix the PCI device capability dict creation
VPD handling was added in ab49f97b2c08294234c7bfd3dedb75780ca519e6 and
introduced a regression which was not caught by unit or functional
tests: if a VPD capability is present for a PCI device, it overrides
the network capabilities obtained from the associated netdev.

This change fixes the issue and adds additional test coverage.

Closes-Bug: #1964995
Change-Id: Ifa9ae93530339828333dd3e23cdffe9f85e96a70