Add nova-emulation to the experimental queue too
This job was moved to periodic in the past to save CI resources. However
sometimes it can be valuable to trigger it on a given review. To be able
to do that easily this patch also adds the job to the experimental
queue.

Then the following gerrit comment can be used to trigger it:
        check experimental

Change-Id: I6a018b57de5afe772c2478aff836b39227dce228