db: Narrow down deprecation warning filter
One of our 'SADeprecationWarning' warning filters is a bit of an odd
duck: unlike all the other filters, this one is applied to all modules
and not just nova. We can't fix issues caused by code that isn't nova
(at least, not in the nova tree) so this is a silly approach. Remove it.

Change-Id: I803d31117d0536df2e436a2f64144e4029c9073c
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>