[stable-only] Drop lower-constraints job
During the PTG the TC discussed the topic and decided to drop the job
completely. Since the latest job configuration broke all stable gate
for nova (older than yoga) this is needed there to unblock our gates.
For dropping the job on master let's wait to the resolution as the
gate is not broken there, hence the patch is stable-only.

Change-Id: I514f6b337ffefef90a0ce9ab0b4afd083caa277e