Switch libvirt event timeout message to warning
At the moment, if libvirt times out in detaching a device, it
reports this as an ERROR even if the process will be retried
and eventually succeed.

We should just log a warning since there's nothing to do, and
if the process fails after all the retries, it will log an ERROR
anyways.

Closes-Bug: #1972023
Change-Id: Idda12db5758706a97b7841571b9ecd3dc6e6905e