Try to look up hosts by hypervisor in agg ops
When adding or removing aggregates, Nova tries to look up the host
with `get_all_by_host`, which looks at the "host" attribute and
notably NOT "hypervisor_hostname." In Ironic, all nodes will be
registered as hypervisors, but all will have the same host value
(e.g., "<fqdn>-ironic".)

When removing hosts by hypervisor hostname, which is how our fork
of Nova does things, due to how Blazar is reserving the nodes, passing
the hypervisor hostname in as the search query will not work, because
it's not looking at this field.

So, this adds a new flag to allow falling back to the hypervisor
hostname in this situation. The function is used in a lot of places
so by default there is no modification to the business logic if the
flag remains at its default value.

N.B.: this also loosens an error detection on _get_service_in_cell_by_host
so that Service.get_by_compute_host can raise a ComputeHostNotFound
exception, and that will be caught and handled by _find_service_in_cell.
When looking up a service by hypervisor, that function will raise this
error, which we need to catch and run the fallback logic to enable
the lookup to succeed.