Make test_wait_for_instance_event_* test time independent
I91c42af8226dc8607d86c52413924351ba0586ac introduced three test cases
that unfortunately dependent on execution time. The code under test uses
a StopWatch  to measure execution time and include that in a log that is
then assert by the test cases. This cases the test assertion of "0.00
seconds" to fail on a slow CI node.

This patch removes the dependency towards StopWatch.elapsed to stabilize
the tests.

Closes-Bug: #1978817
Change-Id: Id1ba3a886cf2d31956fc8f51dc7de08a79b22922