Update Availability zone doc page
Availability zone doc page [1] mentioned that server moving
operation (evacuate/live-migrate) request can force target host
and bypass scheduler.

This support of forcing target host for evacuate/live-migrate
request was removed in [2].

This change updates the doc page by mentioning that it was possible
to force a host with older API microversion and how to use it, if
user wants to force target host for evacuate/live-migrate.

[1] https://docs.openstack.org/nova/latest/admin/availability-zones.html#implications-for-moving-servers
[2] https://review.opendev.org/q/topic:bp%252Fremove-force-flag-from-live-migrate-and-evacuate

Closes-Bug: #1947421
Change-Id: Iaf989ccbde49a8c3735d260b2dd4a319bc77caf8