update nova-next and nova-ovs-hybrid-plug to disable linuxbridge
neutron now declares linuxbridge experimental and requires
it to be enabled by a new config option. We do not set
that in these jobs so the neutron server fails to load
since we enable the linux bridge mech driver.

we do not actully use linuxbridge in these jobs so this patch
just disables the mech driver.

Change-Id: I129a850c3327498d26d899c7709f5749622afdd8
Closes-Bug: #1980948