zuul: Put Centos9 Stream job periodic-weekly and experimental
This patch is a followup on 45c0a3884b6b4c0b9a6b1b3cef47651f5504502e

As agreed on the Nova weekly meeting [1], we won't put back C9S job to voting
and we'll rather put it on both experimental and periodic-weekly pipelines.
In order to insure we will monitor the status of those runs, we'll check
the status of the job during every Nova meeting happening weekly.

In order to trigger potential race conditions, it was acted to provide an
experimental job for C9S too.

[1] https://meetings.opendev.org/meetings/nova/2022/nova.2022-07-05-16.00.log.html#l-52

Change-Id: I2de4d90202fd7e735026150f79409d51f2d70e0e