extend_volume of libvirt/volume/fc should not use device_path
The connection_info['data']['device_path'] field is not always
available. In cases that it was not available, it would cause
the debug code to raise a KeyError instead of proceeding.

Other similar debug messages in the same file do not include
device_path. As a simple fix, just drop the device_path from
the log.

Closes-Bug: #1989894
Change-Id: I4218fb39ce80800be8c95da7132b3b74226ad10a