Test ceph-multistore with a real image
This inflates the cirros image to 1G for a more realistic scenario.
Technically we should have been doing something like this all along,
as the deployment guidance for ceph is to use a raw image, not a qcow2
one, so this also increases our accuracy to real-life.

We also need to up the volume size tempest uses for various tests
to make sure we will fit.

Change-Id: I5c447e630aaf1413a5eac89c2e8103506d245221