Bump nova-ceph-multstore timeout
The recent change(s) to enable a lot more SSHABLE checks puts the
runtime of the ceph job really close to the 2h timeout even when
things are working. Sometimes it times out before it finishes even
though things are progressing. Bump the timeout to avoid that.

Also bump us to 8G swap to match what is set on the parent ceph job
when we upgraded to jammy. We could just unset this, but better to
pin it high in case that job (defined elsewhere) changes. Our job
is the largest ceph job, so it makes sense that it keeps its own
swap level high.

Change-Id: I6cefd87671614d87d92e4675fbc989fc9453c8b9
(cherry picked from commit 6ff3237149c64bd01fb99149744131a15acab52e)