[stable-only] Remove nova-emulation from check pipeline of Yoga
Yoga will transition soon to Unmaintained and nova-emulation job is
almost constantly failing on stable/yoga now, so this patch moves it
to experimental pipeline.
Note that nova-emulation is only running on stable/yoga, and we don't
run it on newer branches. Only periodic-weekly pipeline is targeted for
this job.

Change-Id: Iff6ffab833754cf39c47483a48a5df3b749fcffd